More changes for Mac OS X build. Still doesn't build on Mac OS X yet, but these changes shouldn't cause trouble to anyone else.
Added __APPLE__ section to os.h, which is mirroring the linux section to some
degree.

Mostly added some typecasts that arent' needed on other platforms.

+++ src/trunk/base/config.c
@@ -647,7 +647,7 @@ do_config(struct _como * m, int argc, char *argv[])
 	if (m->maxfilesize > 1024*1024*1024) { 
 	    m->maxfilesize = DEFAULT_FILESIZE; 
 	    sprintf(errstr, "'filesize' should be < 1GB --> set to %dMB\n", 
-		   m->maxfilesize / (1024*1024));
+		    (int)(m->maxfilesize / (1024*1024)));
 	    return errstr; 
 	} 
 	break;
